requirements.txt
beautifulsoup4==4.12.3
pandas==2.0.1
Requests==2.32.3
tqdm==4.65.0
pillow==9.5.0
lxml